demo updates

+++ examples/vertica/demo/setup_vertica.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
 version="$1"
+db_port="$2"
 db_user=redisconnect
 db_pwd=Redis123
 [[ -z "$version" ]] && { echo "Error: Missing docker version tag e.g. 11.1.0-0, latest"; exit 1; }
+[[ -z "$db_port" ]] && { echo "Error: Missing database port e.g. 5433"; exit 1; }
 
-container_name="vertica-$(hostname)"
+container_name="vertica-$(hostname)-$db_port"
 # delete the existing container if it exist
 sudo docker kill $container_name;sudo docker rm $container_name;
 
@@ -16,12 +18,12 @@ sudo chmod a+w -R $(pwd)/$container_name/data
 
 echo "Creating $container_name docker container."
 sudo docker run --name $container_name \
-	-p 5433:5433 \
+	-p $db_port:5433 \
 	-e APP_DB_USER=$db_user \
 	-e APP_DB_PASSWORD=$db_pwd \
 	-e VERTICA_DB_NAME="RedisConnect" \
 	-v $(pwd)/$container_name/data:/data \
-  -d vertica/vertica-ce:$version
+        -d vertica/vertica-ce:$version
 
 while ! nc -vz $(sudo docker inspect --format='{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}' $container_name) 5433 < /dev/null
 do
